analytics_service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=2.0.0
pandas>=2.3.0
scikit-learn>=1.7.0
scipy>=1.16.0
xgboost>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0


